Refactoring: Reuse code responsible for implementing mapping through initialization block
