+++ MappingGenerator/MappingGenerator/MappingGenerator/MappingHelper.cs
@@ -1,7 +1,9 @@
 using System;
+using System.Linq;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
+using Microsoft.CodeAnalysis.Editing;
 
 namespace MappingGenerator
 {
@@ -38,6 +40,31 @@ namespace MappingGenerator
         {
             return initializerExpressionSyntax
                 .WithLeadingTrivia(SyntaxTriviaList.Create(SyntaxFactory.EndOfLine(Environment.NewLine)));
+        }
+
+        public static ObjectCreationExpressionSyntax AddInitializerWithMapping(
+            this ObjectCreationExpressionSyntax objectCreationExpression, IMappingSourceFinder mappingSourceFinder,
+            ITypeSymbol createdObjectTyp)
+        {
+            var propertiesToSet = ObjectHelper.GetPublicPropertySymbols(createdObjectTyp).Where(x => x.SetMethod?.DeclaredAccessibility == Accessibility.Public);
+            var assigments =  propertiesToSet.Select(x =>
+            {
+                var src = mappingSourceFinder.FindMappingSource(x.Name, x.Type);
+                if (src != null)
+                {
+                    return SyntaxFactory.AssignmentExpression(SyntaxKind.SimpleAssignmentExpression, SyntaxFactory.IdentifierName(x.Name), src.Expression);
+                }
+
+                return null;
+            }).OfType<ExpressionSyntax>();
+
+            var initializerExpressionSyntax = SyntaxFactory.InitializerExpression(SyntaxKind.ObjectInitializerExpression,new SeparatedSyntaxList<ExpressionSyntax>().AddRange(assigments)).FixInitializerExpressionFormatting();
+            return objectCreationExpression.WithInitializer(initializerExpressionSyntax);
+        }
+
+        public static ObjectCreationExpressionSyntax CreateObjectCreationExpressionWithInitializer(ITypeSymbol targetType, IMappingSourceFinder subMappingSourceFinder, SyntaxGenerator syntaxGenerator, SemanticModel semanticModel)
+        {
+            return ((ObjectCreationExpressionSyntax) syntaxGenerator.ObjectCreationExpression(targetType)).AddInitializerWithMapping(subMappingSourceFinder, targetType);
         }
     }
 }
